+++ src/vite/bytecode/index.ts
@@ -7,7 +7,7 @@ import path from 'node:path'
 import { createFilter, normalizePath } from 'vite'
 
 import { bytecodeId, bytecodeLog } from '../constant'
-import { readableSize } from '../utils'
+import { readableSize } from '../utils/file'
 import { bytecodeModuleLoaderCode } from './code'
 import {
   bytecodeModuleLoader,
@@ -18,20 +18,23 @@ import {
   useStrict,
 } from './utils'
 
+/** Options for bytecode compilation */
 export interface BytecodeOptions {
+  /** Enable bytecode compilation */
   enable: boolean
   /**
    * Enable in preload script. Remember to set `sandbox: false` when creating window
    */
   preload?: boolean
   /**
-   * Custom electron binary path
+   * Custom electron binary path for bytecode generation
    */
   electronPath?: string
   /**
    * Before transformed code compile function. If return `Falsy` value, it will be ignored
-   * @param code transformed code
-   * @param id file path
+   * @param code - Transformed code
+   * @param id - File path
+   * @returns Transformed code or falsy to ignore
    */
   beforeCompile?: (code: string, id: string) => Promisable<string | null | undefined | void>
 }
@@ -41,7 +44,15 @@ function getBytecodeLoaderBlock(chunkFileName: string): string {
 }
 
 /**
- * Compile to v8 bytecode to protect source code.
+ * Vite plugin that compiles JavaScript files to V8 bytecode for source code protection.
+ *
+ * This plugin transforms JavaScript files into bytecode format during build,
+ * making the source code more difficult to reverse engineer. The generated bytecode
+ * files (.jsc) can only be executed in compatible Electron environments.
+ *
+ * @param env - Environment type ('preload' | 'main')
+ * @param options - Bytecode compilation options
+ * @returns Vite plugin or null if bytecode compilation is disabled
  */
 export function bytecodePlugin(env: 'preload' | 'main', options: BytecodeOptions): Plugin | null {
   const { enable, preload = false, electronPath, beforeCompile } = options
