+++ src/vite/utils/key.ts
@@ -2,26 +2,50 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { generate } from 'selfsigned'
 
-import { log } from './constant'
+import { log } from '../constant'
 
+/**
+ * Distinguished name for certificate generation
+ */
 export interface DistinguishedName {
+  /** Country name (2-letter code) */
   countryName?: string
+  /** State or province name */
   stateOrProvinceName?: string
+  /** Locality/city name */
   localityName?: string
+  /** Organization name */
   organizationName?: string
+  /** Organizational unit name */
   organizationalUnitName?: string
+  /** Common name */
   commonName?: string
+  /** Serial number */
   serialNumber?: string
+  /** Title */
   title?: string
+  /** Description */
   description?: string
+  /** Business category */
   businessCategory?: string
+  /** Email address */
   emailAddress?: string
 }
+
+/** Certificate subject name-value pairs */
 export type CertSubject = {
   name: string
   value: string
 }[]
 
+/**
+ * Generate a new key pair for signing
+ * @param keyLength - Key length in bits
+ * @param subject - Certificate subject
+ * @param days - Validity period in days
+ * @param privateKeyPath - Output path for private key
+ * @param certPath - Output path for certificate
+ */
 export async function generateKeyPair(
   keyLength: number,
   subject: CertSubject,
@@ -54,14 +78,25 @@ export async function generateKeyPair(
   fs.writeFileSync(certPath, cert.replace(/\r\n?/g, '\n'))
 }
 
+/** Options for parsing keys */
 export interface GetKeysOption {
+  /** Path to private key file */
   privateKeyPath: string
+  /** Path to certificate file */
   certPath: string
+  /** Key length for generation */
   keyLength: number
+  /** Certificate subject */
   subject: DistinguishedName
+  /** Validity period in days */
   days: number
 }
 
+/**
+ * Parse and load keys, generating new ones if they don't exist
+ * @param options - Key parsing options
+ * @returns Object containing private key and certificate strings
+ */
 export async function parseKeys({
   keyLength,
   privateKeyPath,
@@ -93,6 +128,11 @@ export async function parseKeys({
   return { privateKey, cert }
 }
 
+/**
+ * Convert DistinguishedName object to CertSubject array
+ * @param subject - Distinguished name object
+ * @returns Certificate subject array
+ */
 function parseSubjects(subject: DistinguishedName): CertSubject {
   return Object.entries(subject)
     .filter(([_, value]) => !!value)
